feat: add use case to viewmodel and fix height data type

+++ app/src/main/java/com/example/nutrition_assessment_system_android_app/data/user/datasource/remote/request/UpdateUserProfileRequest.kt
@@ -4,7 +4,7 @@ data class UpdateUserProfileRequest(
     val gender: String,
     val age: Int,
     val weight: Float,
-    val height: Float,
+    val height: Int,
     val activityLevel: String,
     val goal: String,
 )
