+++ app/src/main/java/com/example/nutrition_assessment_system_android_app/domain/params/UpdateUserProfileParams.kt
@@ -3,7 +3,7 @@ package com.example.nutrition_assessment_system_android_app.domain.params
 data class UpdateUserProfileParams(
     val gender: String,
     val age: Int,
-    val height: Float,
+    val height: Int,
     val weight: Float,
     val activityLevel: String,
     val goal: String
